requirements.txt
flask==2.0.1
streamlit==1.0.0
transformers==4.11.3
torch==1.9.0
openai==0.9.0
pyyaml==5.4.1
requests==2.26.0
